Fixed a lot of warnings to make the build cleaner, mostly type casts and vector iterators.

+++ trunk/Robot/controller/Camera.cpp
@@ -263,7 +263,7 @@ void Camera::Update()
 		return;
 
 	time_t seconds = time(NULL);
-	float interval = seconds - lastTime_;
+	float interval = (float)seconds - lastTime_;
 	lastTime_ = seconds;
 
 	scanTimer_ += interval;
